robot cleanup and tests!
