+++ src/models/tests/gain.test.ts
@@ -1,0 +1,106 @@
+import { MockAudioContext, mockBody } from "./mock";
+mockBody();
+
+import { globalZoneID } from "../../config";
+import { User } from "../user";
+
+describe("User gain tests", () => {
+  test("Gain: speaker on the max right of listener", () => {
+    const audioCtx = new MockAudioContext();
+    window.audioContext = audioCtx;
+
+    const local = new User("local", null, 100, 100, true);
+    local["earshotDistance"] = 100;
+    local["zoneData"].zoneID = globalZoneID;
+
+    const remote = new User("remote", null, 200, 100);
+    remote.media["audioTrack"] = new MediaStreamTrack();
+    remote["zoneData"].zoneID = globalZoneID;
+
+    const wantGain = 0;
+    const audioMod = local["getAudioMod"](100, remote.getPos());
+    expect(audioMod.gain).toBe(wantGain);
+
+    local.processUsers([remote]);
+
+    const nodeChain = remote.media["nodeChain"];
+    expect(nodeChain.getGain()).toBe(wantGain);
+  });
+
+  test("Gain: speaker on the max left of listener", () => {
+    const audioCtx = new MockAudioContext();
+    window.audioContext = audioCtx;
+
+    const local = new User("local", null, 100, 100, true);
+    local["earshotDistance"] = 100;
+
+    const remote = new User("remote", null, 0, 100);
+    remote.media["audioTrack"] = new MediaStreamTrack();
+
+    const wantGain = 0;
+    const audioMod = local["getAudioMod"](100, remote.getPos());
+    expect(audioMod.gain).toBe(wantGain);
+
+    local.processUsers([remote]);
+
+    const nodeChain = remote.media["nodeChain"];
+    expect(nodeChain.getGain()).toBe(wantGain);
+  });
+
+  test("Gain: speaker on the partial right of listener", () => {
+    const audioCtx = new MockAudioContext();
+    window.audioContext = audioCtx;
+
+    const local = new User("local", null, 100, 100, true);
+    local["earshotDistance"] = 100;
+
+    const remote = new User("remote", null, 150, 100);
+    remote.media["audioTrack"] = new MediaStreamTrack();
+
+    const wantGain = 0.25;
+    const audioMod = local["getAudioMod"](50, remote.getPos());
+    expect(audioMod.gain).toBe(wantGain);
+
+    local.processUsers([remote]);
+    const nodeChain = remote.media["nodeChain"];
+    expect(nodeChain.getGain()).toBe(wantGain);
+  });
+
+  test("Gain: speaker on the partial left of listener", () => {
+    const audioCtx = new MockAudioContext();
+    window.audioContext = audioCtx;
+
+    const local = new User("local", null, 100, 100, true);
+    local["earshotDistance"] = 100;
+
+    const remote = new User("remote", null, 50, 100);
+    remote.media["audioTrack"] = new MediaStreamTrack();
+
+    const wantGain = 0.25;
+    const audioMod = local["getAudioMod"](50, remote.getPos());
+    expect(audioMod.gain).toBe(wantGain);
+
+    local.processUsers([remote]);
+    const nodeChain = remote.media["nodeChain"];
+    expect(nodeChain.getGain()).toBe(wantGain);
+  });
+
+  test("Gain: speaker in the center of the listener", () => {
+    const audioCtx = new MockAudioContext();
+    window.audioContext = audioCtx;
+
+    const local = new User("local", null, 100, 100, true);
+    local["earshotDistance"] = 100;
+
+    const remote = new User("remote", null, 100, 100);
+    remote.media["audioTrack"] = new MediaStreamTrack();
+
+    const wantGain = 0.5;
+    const audioMod = local["getAudioMod"](0, remote.getPos());
+    expect(audioMod.gain).toBe(wantGain);
+
+    local.processUsers([remote]);
+    const nodeChain = remote.media["nodeChain"];
+    expect(nodeChain.getGain()).toBe(wantGain);
+  });
+});
